Move robot load things to common area
