+++ src/mx_bluesky/hyperion/experiment_plans/robot_load_and_change_energy.py
@@ -10,7 +10,7 @@ import bluesky.preprocessors as bpp
 import pydantic
 from blueapi.core import BlueskyContext
 from bluesky.utils import Msg
-from dodal.devices.aperturescatterguard import ApertureScatterguard, ApertureValue
+from dodal.devices.aperturescatterguard import ApertureScatterguard
 from dodal.devices.attenuator.attenuator import BinaryFilterAttenuator
 from dodal.devices.backlight import Backlight, BacklightPosition
 from dodal.devices.focusing_mirror import FocusingMirrorWithStripes, MirrorVoltages
@@ -19,13 +19,16 @@ from dodal.devices.i03.undulator_dcm import UndulatorDCM
 from dodal.devices.motors import XYZPositioner
 from dodal.devices.oav.oav_detector import OAV
 from dodal.devices.robot import BartRobot, SampleLocation
-from dodal.devices.smargon import Smargon, StubPosition
+from dodal.devices.smargon import Smargon
 from dodal.devices.thawer import Thawer
 from dodal.devices.webcam import Webcam
 from dodal.devices.xbpm_feedback import XBPMFeedback
-from dodal.plan_stubs.motor_utils import MoveTooLarge, home_and_reset_wrapper
 
-from mx_bluesky.common.utils.log import LOGGER
+from mx_bluesky.common.device_setup_plans.robot_load_unload import (
+    do_plan_while_lower_gonio_at_home,
+    prepare_for_robot_load,
+    wait_for_smargon_not_disabled,
+)
 from mx_bluesky.hyperion.experiment_plans.set_energy_plan import (
     SetEnergyComposite,
     set_energy_plan,
@@ -61,25 +64,6 @@ def create_devices(context: BlueskyContext) -> RobotLoadAndEnergyChangeComposite
     return device_composite_from_context(context, RobotLoadAndEnergyChangeComposite)
 
 
-def wait_for_smargon_not_disabled(smargon: Smargon, timeout=60):
-    """Waits for the smargon disabled flag to go low. The robot hardware is responsible
-    for setting this to low when it is safe to move. It does this through a physical
-    connection between the robot and the smargon.
-    """
-    LOGGER.info("Waiting for smargon enabled")
-    SLEEP_PER_CHECK = 0.1
-    times_to_check = int(timeout / SLEEP_PER_CHECK)
-    for _ in range(times_to_check):
-        smargon_disabled = yield from bps.rd(smargon.disabled)
-        if not smargon_disabled:
-            LOGGER.info("Smargon now enabled")
-            return
-        yield from bps.sleep(SLEEP_PER_CHECK)
-    raise TimeoutError(
-        "Timed out waiting for smargon to become enabled after robot load"
-    )
-
-
 def take_robot_snapshots(oav: OAV, webcam: Webcam, directory: Path):
     time_now = datetime.now()
     snapshot_format = f"{time_now.strftime('%H%M%S')}_{{device}}_after_load"
@@ -91,31 +75,6 @@ def take_robot_snapshots(oav: OAV, webcam: Webcam, directory: Path):
         # Note: should be able to use `wait=True` after https://github.com/bluesky/bluesky/issues/1795
         yield from bps.trigger(device, group="snapshots")
         yield from bps.wait("snapshots")
-
-
-def prepare_for_robot_load(
-    aperture_scatterguard: ApertureScatterguard, smargon: Smargon
-):
-    yield from bps.abs_set(
-        aperture_scatterguard.selected_aperture,
-        ApertureValue.OUT_OF_BEAM,
-        group="prepare_robot_load",
-    )
-
-    yield from bps.mv(smargon.stub_offsets, StubPosition.RESET_TO_ROBOT_LOAD)
-
-    # fmt: off
-    yield from bps.mv(
-        smargon.x, 0,
-        smargon.y, 0,
-        smargon.z, 0,
-        smargon.omega, 0,
-        smargon.chi, 0,
-        smargon.phi, 0
-    )
-    # fmt: on
-
-    yield from bps.wait("prepare_robot_load")
 
 
 def do_robot_load(
@@ -140,16 +99,6 @@ def do_robot_load(
         group="thawing_finished",
     )
     yield from wait_for_smargon_not_disabled(composite.smargon)
-
-
-def raise_exception_if_moved_out_of_cryojet(exception):
-    yield from bps.null()
-    if isinstance(exception, MoveTooLarge):
-        raise Exception(
-            f"Moving {exception.axis} back to {exception.position} after \
-                        robot load would move it out of the cryojet. The max safe \
-                        distance is {exception.maximum_move}"
-        )
 
 
 def pin_already_loaded(
@@ -179,22 +128,9 @@ def robot_load_and_snapshots(
         thawing_time,
     )
 
-    # The lower gonio must be in the correct position for the robot load and we
-    # want to put it back afterwards. Note we don't wait the robot is interlocked
-    # to the lower gonio and the  move is quicker than the robot takes to get to the
-    # load position.
-    yield from bpp.contingency_wrapper(
-        home_and_reset_wrapper(
-            robot_load_plan,
-            composite.lower_gonio,
-            BartRobot.LOAD_TOLERANCE_MM,
-            CONST.HARDWARE.CRYOJET_MARGIN_MM,
-            "lower_gonio",
-            wait_for_all=False,
-        ),
-        except_plan=raise_exception_if_moved_out_of_cryojet,
+    gonio_finished = yield from do_plan_while_lower_gonio_at_home(
+        robot_load_plan, composite.lower_gonio
     )
-
     yield from bps.wait(group="snapshot")
 
     yield from take_robot_snapshots(composite.oav, composite.webcam, snapshot_directory)
@@ -205,7 +141,7 @@ def robot_load_and_snapshots(
     yield from bps.read(composite.webcam)
     yield from bps.save()
 
-    yield from bps.wait("reset-lower_gonio")
+    yield from bps.wait(gonio_finished)
 
 
 def robot_load_and_change_energy_plan(
